% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.polygonsPRISM.R
\name{average.polygonsPRISM}
\alias{average.polygonsPRISM}
\title{Calculate polygons average weather variable from PRISM raster}
\usage{
average.polygonsPRISM(varName, dayStart, dayEnd, data_dir, shpFile)
}
\arguments{
\item{varName}{Variable name is either 'ppt', 'tmax', or 'tmin'.}

\item{dayStart}{Initial day must be in format 'Y/m/d'}

\item{dayEnd}{Last day must be in format 'Y/m/d'}

\item{data_dir}{Directory where prism data is}

\item{shpFile}{Polygons shapefile in "SpatialPolygonsDataFrame".}
}
\value{
Average mean of squared grid within the polygon for either variable.
}
\description{
TBD
}
