% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.polygonsPRISM.R
\name{averge.polygonsPRISM}
\alias{averge.polygonsPRISM}
\title{Calculate polygons average weather variable from PRISM raster}
\usage{
averge.polygonsPRISM(varName, day, data_dir, shpFile)
}
\arguments{
\item{varName}{Variable name is either 'ppt', 'tmax', or 'tmin'.}

\item{day}{Day must be in format yyyymmdd (e.g., 199910113, for yyyy = 1991, mm = 10, and dd = 13).}

\item{data_dir}{Directory where prism data is}

\item{shpFile}{Polygons shapefile in "SpatialPolygonsDataFrame".}
}
\value{
Average mean of squared grid within the polygon for either variable.
}
\description{
Calculate polygons average weather variable from PRISM raster
}
