% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd.calculation.R
\name{gdd.calculation}
\alias{gdd.calculation}
\title{Growing Degree Days (GDD) calculation.}
\usage{
gdd.calculation(tmin, tmax, tbase)
}
\arguments{
\item{tmin}{Minimum temperature}

\item{tmax}{Maximum temperature}

\item{tbase}{Crop specific base temperature}
}
\value{
GDD
}
\description{
gdd.calculation() calculates crop specific GDD using as daily average the formula 0.5*(tmin + tmax). tmin and tmax must be vectors and have the same dimensions. tbase must be a scalar.
}
