% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.polygonsPRISM.R
\name{averge.polygonsPRISM}
\alias{averge.polygonsPRISM}
\title{Calculate polygons average weather variable from PRISM raster}
\usage{
averge.polygonsPRISM(varName, dayStart, dayEnd, data_dir, shpFile)
}
\arguments{
\item{varName}{Variable name is either 'ppt', 'tmax', or 'tmin'.}

\item{dayStart}{Initial day must be in format %Y/%m/%d  (e.g., 19991/01/13, for %Y = 1991, %m = 10, and %d = 13).}

\item{dayEnd}{Last day must be in format %Y/%m/%d  (e.g., 19991/01/13, for %Y = 1991, %m = 10, and %d = 13).}

\item{data_dir}{Directory where prism data is}

\item{shpFile}{Polygons shapefile in "SpatialPolygonsDataFrame".}
}
\value{
Average mean of squared grid within the polygon for either variable.
}
\description{
Calculate polygons average weather variable from PRISM raster
}
